Testing

+++ c_lib/Stepper.h
@@ -20,7 +20,7 @@ typedef struct {
    int motor_pin_1; // indicate which pins the motors are on - we only need one
    int step_number; // what number of steps the motor is currently on
    float last_step_time;
-   float step_delay; // parameter to control speed
+   double step_delay; // parameter to control speed
 } Stepper_t;
 
 /**
